requirements.txt
Flask==2.3.2
google-cloud-firestore==2.11.0
openai==0.28.0
pydantic==1.10.15
python-dotenv==0.21.1
functions-framework==3.4.1
uuid==1.30
